Safari-611.1.10/LayoutTests/inspector/runtime/callFunctionOn-emulateUserGesture-expected.txt

Tests for Runtime.callFunctionOn emulateUserGesture option.


== Running test suite: Runtime.callFunctionOn.emulateUserGesture
-- Running test case: CallFunctionOnWithoutEmulatingUserGesture
Not in User Gesture
-- Running test case: CallFunctionOnWithEmulatingUserGesture
In User Gesture
